+++ Source/DataSources/KmlDataSource.js
@@ -43,6 +43,7 @@ import Autolinker from "../ThirdParty/Autolinker.js";
 import Uri from "../ThirdParty/Uri.js";
 import when from "../ThirdParty/when.js";
 import zip from "../ThirdParty/zip.js";
+import getElement from "../Widgets/getElement.js";
 import BillboardGraphics from "./BillboardGraphics.js";
 import CompositePositionProperty from "./CompositePositionProperty.js";
 import DataSource from "./DataSource.js";
@@ -182,7 +183,7 @@ var featureTypes = {
   NetworkLink: processNetworkLink,
   GroundOverlay: processGroundOverlay,
   PhotoOverlay: processUnsupportedFeature,
-  ScreenOverlay: processUnsupportedFeature,
+  ScreenOverlay: processScreenOverlay,
   Tour: processTour,
 };
 
@@ -2490,6 +2491,160 @@ function processLookAt(featureNode, entity, ellipsoid) {
   }
 }
 
+function processScreenOverlay(
+  dataSource,
+  screenOverlayNode,
+  processingData,
+  deferredLoading
+) {
+  var screenOverlay = processingData.screenOverlayContainer;
+  if (!defined(screenOverlay)) {
+    return undefined;
+  }
+
+  var sourceResource = processingData.sourceResource;
+  var uriResolver = processingData.uriResolver;
+
+  var iconNode = queryFirstNode(screenOverlayNode, "Icon", namespaces.kml);
+  var icon = getIconHref(
+    iconNode,
+    dataSource,
+    sourceResource,
+    uriResolver,
+    false
+  );
+
+  if (!defined(icon)) {
+    return undefined;
+  }
+
+  var img = document.createElement("img");
+  dataSource._screenOverlays.push(img);
+
+  img.src = icon.url;
+  img.onload = function () {
+    var styles = ["position: absolute"];
+
+    var screenXY = queryFirstNode(
+      screenOverlayNode,
+      "screenXY",
+      namespaces.kml
+    );
+    var overlayXY = queryFirstNode(
+      screenOverlayNode,
+      "overlayXY",
+      namespaces.kml
+    );
+    var size = queryFirstNode(screenOverlayNode, "size", namespaces.kml);
+
+    var x, y;
+    var xUnit, yUnit;
+    var xStyle, yStyle;
+
+    if (defined(size)) {
+      x = queryNumericAttribute(size, "x");
+      y = queryNumericAttribute(size, "y");
+      xUnit = queryStringAttribute(size, "xunits");
+      yUnit = queryStringAttribute(size, "yunits");
+
+      if (defined(x) && x !== -1 && x !== 0) {
+        if (xUnit === "fraction") {
+          xStyle = "width: " + Math.floor(x * 100) + "%";
+        } else if (xUnit === "pixels") {
+          xStyle = "width: " + x + "px";
+        }
+
+        styles.push(xStyle);
+      }
+
+      if (defined(y) && y !== -1 && y !== 0) {
+        if (yUnit === "fraction") {
+          yStyle = "height: " + Math.floor(y * 100) + "%";
+        } else if (yUnit === "pixels") {
+          yStyle = "height: " + y + "px";
+        }
+
+        styles.push(yStyle);
+      }
+    }
+
+    // set the interim style so the width/height properties get calculated
+    img.style = styles.join(";");
+
+    var xOrigin = 0;
+    var yOrigin = img.height;
+
+    if (defined(overlayXY)) {
+      x = queryNumericAttribute(overlayXY, "x");
+      y = queryNumericAttribute(overlayXY, "y");
+      xUnit = queryStringAttribute(overlayXY, "xunits");
+      yUnit = queryStringAttribute(overlayXY, "yunits");
+
+      if (defined(x)) {
+        if (xUnit === "fraction") {
+          xOrigin = x * img.width;
+        } else if (xUnit === "pixels") {
+          xOrigin = x;
+        } else if (xUnit === "insetPixels") {
+          xOrigin = x;
+        }
+      }
+
+      if (defined(y)) {
+        if (yUnit === "fraction") {
+          yOrigin = y * img.height;
+        } else if (yUnit === "pixels") {
+          yOrigin = y;
+        } else if (yUnit === "insetPixels") {
+          yOrigin = y;
+        }
+      }
+    }
+
+    if (defined(screenXY)) {
+      x = queryNumericAttribute(screenXY, "x");
+      y = queryNumericAttribute(screenXY, "y");
+      xUnit = queryStringAttribute(screenXY, "xunits");
+      yUnit = queryStringAttribute(screenXY, "yunits");
+
+      if (defined(x)) {
+        if (xUnit === "fraction") {
+          xStyle =
+            "left: " + "calc(" + Math.floor(x * 100) + "% - " + xOrigin + "px)";
+        } else if (xUnit === "pixels") {
+          xStyle = "left: " + (x - xOrigin) + "px";
+        } else if (xUnit === "insetPixels") {
+          xStyle = "right: " + (x - xOrigin) + "px";
+        }
+
+        styles.push(xStyle);
+      }
+
+      if (defined(y)) {
+        if (yUnit === "fraction") {
+          yStyle =
+            "bottom: " +
+            "calc(" +
+            Math.floor(y * 100) +
+            "% - " +
+            yOrigin +
+            "px)";
+        } else if (yUnit === "pixels") {
+          yStyle = "bottom: " + (y - yOrigin) + "px";
+        } else if (yUnit === "insetPixels") {
+          yStyle = "top: " + (y - yOrigin) + "px";
+        }
+
+        styles.push(yStyle);
+      }
+    }
+
+    img.style = styles.join(";");
+  };
+
+  screenOverlay.appendChild(img);
+}
+
 function processGroundOverlay(
   dataSource,
   groundOverlay,
@@ -2936,6 +3091,7 @@ function processNetworkLink(dataSource, node, processingData, deferredLoading) {
         sourceUri: newSourceUri,
         uriResolver: uriResolver,
         context: networkEntity.id,
+        screenOverlayContainer: processingData.screenOverlayContainer,
       };
       var networkLinkCollection = new EntityCollection();
       var promise = load(dataSource, networkLinkCollection, href, options)
@@ -3101,6 +3257,7 @@ function loadKml(
   kml,
   sourceResource,
   uriResolver,
+  screenOverlayContainer,
   context
 ) {
   entityCollection.removeAll();
@@ -3153,6 +3310,7 @@ function loadKml(
         sourceResource: sourceResource,
         uriResolver: uriResolver,
         context: context,
+        screenOverlayContainer: screenOverlayContainer,
       };
 
       entityCollection.suspendEvents();
@@ -3165,7 +3323,13 @@ function loadKml(
     });
 }
 
-function loadKmz(dataSource, entityCollection, blob, sourceResource) {
+function loadKmz(
+  dataSource,
+  entityCollection,
+  blob,
+  sourceResource,
+  screenOverlayContainer
+) {
   var zWorkerUrl = buildModuleUrl("ThirdParty/Workers/z-worker-pako.js");
   zip.configure({
     workerScripts: {
@@ -3173,6 +3337,7 @@ function loadKmz(dataSource, entityCollection, blob, sourceResource) {
       inflate: [zWorkerUrl, "./pako_inflate.min.js"],
     },
   });
+
   var reader = new zip.ZipReader(new zip.BlobReader(blob));
   return when(reader.getEntries()).then(function (entries) {
     var promises = [];
@@ -3216,7 +3381,8 @@ function loadKmz(dataSource, entityCollection, blob, sourceResource) {
         entityCollection,
         uriResolver.kml,
         sourceResource,
-        uriResolver
+        uriResolver,
+        screenOverlayContainer
       );
     });
   });
@@ -3227,6 +3393,7 @@ function load(dataSource, entityCollection, data, options) {
   var sourceUri = options.sourceUri;
   var uriResolver = options.uriResolver;
   var context = options.context;
+  var screenOverlayContainer = options.screenOverlayContainer;
 
   var promise = data;
   if (typeof data === "string" || data instanceof Resource) {
@@ -3249,12 +3416,22 @@ function load(dataSource, entityCollection, data, options) {
 
   sourceUri = Resource.createIfNeeded(sourceUri);
 
+  if (defined(screenOverlayContainer)) {
+    screenOverlayContainer = getElement(screenOverlayContainer);
+  }
+
   return when(promise)
     .then(function (dataToLoad) {
       if (dataToLoad instanceof Blob) {
         return isZipFile(dataToLoad).then(function (isZip) {
           if (isZip) {
-            return loadKmz(dataSource, entityCollection, dataToLoad, sourceUri);
+            return loadKmz(
+              dataSource,
+              entityCollection,
+              dataToLoad,
+              sourceUri,
+              screenOverlayContainer
+            );
           }
           return readBlobAsText(dataToLoad).then(function (text) {
             //There's no official way to validate if a parse was successful.
@@ -3301,6 +3478,7 @@ function load(dataSource, entityCollection, data, options) {
               kml,
               sourceUri,
               uriResolver,
+              screenOverlayContainer,
               context
             );
           });
@@ -3312,6 +3490,7 @@ function load(dataSource, entityCollection, data, options) {
         dataToLoad,
         sourceUri,
         uriResolver,
+        screenOverlayContainer,
         context
       );
     })
@@ -3333,6 +3512,7 @@ function load(dataSource, entityCollection, data, options) {
  * @property {Boolean} [clampToGround=false] true if we want the geometry features (Polygons, LineStrings and LinearRings) clamped to the ground.
  * @property {Ellipsoid} [ellipsoid=Ellipsoid.WGS84] The global ellipsoid used for geographical calculations.
  * @property {Credit|String} [credit] A credit for the data source, which is displayed on the canvas.
+ * @property {Element|String} [screenOverlayContainer] A container for ScreenOverlay images.
  */
 
 /**
@@ -3427,6 +3607,8 @@ function KmlDataSource(options) {
   this._resourceCredits = [];
 
   this._kmlTours = [];
+
+  this._screenOverlays = [];
 }
 
 /**
@@ -3606,6 +3788,7 @@ Object.defineProperties(KmlDataSource.prototype, {
  * @param {Resource|String} [options.sourceUri] Overrides the url to use for resolving relative links and other KML network features.
  * @param {Boolean} [options.clampToGround=false] true if we want the geometry features (Polygons, LineStrings and LinearRings) clamped to the ground. If true, lines will use corridors so use Entity.corridor instead of Entity.polyline.
  * @param {Ellipsoid} [options.ellipsoid=Ellipsoid.WGS84] The global ellipsoid used for geographical calculations.
+ * @param {Element|String} [options.screenOverlayContainer] A container for ScreenOverlay images.
  *
  * @returns {Promise.<KmlDataSource>} A promise that will resolve to this instances once the KML is loaded.
  */
@@ -3689,6 +3872,16 @@ KmlDataSource.prototype.load = function (data, options) {
       console.log(error);
       return when.reject(error);
     });
+};
+
+/**
+ * Cleans up any non-entity elements created by the data source. Currently this only affects ScreenOverlay elements.
+ */
+KmlDataSource.prototype.destroy = function () {
+  while (this._screenOverlays.length > 0) {
+    var elem = this._screenOverlays.pop();
+    elem.remove();
+  }
 };
 
 function mergeAvailabilityWithParent(child) {
@@ -3972,7 +4165,9 @@ KmlDataSource.prototype.update = function (time) {
           ellipsoid
         );
 
-        load(that, newEntityCollection, href, { context: entity.id })
+        load(that, newEntityCollection, href, {
+          context: entity.id,
+        })
           .then(
             getNetworkLinkUpdateCallback(
               that,