+++ Source/Scene/ModelExperimental/ModelFeatureTable.js
@@ -1,14 +1,17 @@
 import BatchTexture from "../BatchTexture.js";
+import Cesium3DTileFeature from "../Cesium3DTileFeature.js";
+import Check from "../../Core/Check.js";
+import defined from "../../Core/defined.js";
 import destroyObject from "../../Core/destroyObject.js";
 import ModelFeature from "./ModelFeature.js";
 
 /**
  * Manages the {@link ModelFeature}s in a {@link ModelExperimental}.
- * Extracts the properties from a {@link FeatureTable}.
+ * Extracts the properties from a {@link PropertyTable}.
  *
  * @param {Object} options An object containing the following options:
  * @param {ModelExperimental} options.model The model that owns this feature table.
- * @param {FeatureTable} options.featureTable The feature table from the model used to initialize the model.
+ * @param {PropertyTable} options.propertyTable The property table from the model used to initialize the model.
  *
  * @alias ModelFeatureTable
  * @constructor
@@ -17,8 +20,17 @@ import ModelFeature from "./ModelFeature.js";
  * @experimental This feature is using part of the 3D Tiles spec that is not final and is subject to change without Cesium's standard deprecation policy.
  */
 export default function ModelFeatureTable(options) {
-  this._featureTable = options.featureTable;
-  this._model = options.model;
+  var model = options.model;
+  var propertyTable = options.propertyTable;
+
+  //>>includeStart('debug', pragmas.debug);
+  Check.typeOf.object("propertyTable", propertyTable);
+  Check.typeOf.object("model", model);
+  //>>includeEnd('debug');
+
+  this._propertyTable = propertyTable;
+  this._model = model;
+
   this._features = undefined;
   this._featuresLength = 0;
 
@@ -62,26 +74,36 @@ Object.defineProperties(ModelFeatureTable.prototype, {
 });
 
 function initialize(modelFeatureTable) {
-  var featuresLength = modelFeatureTable._featureTable.count;
+  var content = modelFeatureTable._model.content;
+  var hasContent = defined(content);
+
+  var featuresLength = modelFeatureTable._propertyTable.count;
   if (featuresLength === 0) {
     return;
   }
 
   var features = new Array(featuresLength);
   for (var i = 0; i < featuresLength; i++) {
-    features[i] = new ModelFeature({
-      model: modelFeatureTable._model,
-      featureId: i,
-      featureTable: modelFeatureTable,
-    });
+    if (hasContent) {
+      features[i] = new Cesium3DTileFeature(content, i);
+    } else {
+      features[i] = new ModelFeature({
+        model: modelFeatureTable._model,
+        featureId: i,
+        featureTable: modelFeatureTable,
+      });
+    }
   }
 
   modelFeatureTable._features = features;
   modelFeatureTable._featuresLength = featuresLength;
+
   modelFeatureTable._batchTexture = new BatchTexture({
     featuresLength: featuresLength,
     owner: modelFeatureTable,
-    statistics: modelFeatureTable._statistics,
+    statistics: hasContent
+      ? content.tileset.statistics
+      : modelFeatureTable._statistics,
   });
 }
 
@@ -96,28 +118,52 @@ ModelFeatureTable.prototype.update = function (frameState) {
   this._batchTexture.update(undefined, frameState);
 };
 
+ModelFeatureTable.prototype.setShow = function (featureId, show) {
+  this._batchTexture.setShow(featureId, show);
+};
+
+ModelFeatureTable.prototype.setAllShow = function (show) {
+  this._batchTexture.setAllShow(show);
+};
+
+ModelFeatureTable.prototype.getShow = function (featureId) {
+  return this._batchTexture.getShow(featureId);
+};
+
+ModelFeatureTable.prototype.setColor = function (featureId, color) {
+  this._batchTexture.setColor(featureId, color);
+};
+
+ModelFeatureTable.prototype.setAllColor = function (color) {
+  this._batchTexture.setAllColor(color);
+};
+
+ModelFeatureTable.prototype.getColor = function (featureId, result) {
+  return this._batchTexture.getColor(featureId, result);
+};
+
+ModelFeatureTable.prototype.getPickColor = function (featureId) {
+  return this._batchTexture.getPickColor(featureId);
+};
+
 ModelFeatureTable.prototype.getFeature = function (featureId) {
   return this._features[featureId];
 };
 
 ModelFeatureTable.prototype.hasProperty = function (featureId, propertyName) {
-  return this._featureTable.hasProperty(featureId, propertyName);
+  return this._propertyTable.hasProperty(featureId, propertyName);
 };
 
 ModelFeatureTable.prototype.getProperty = function (featureId, name) {
-  return this._featureTable.getProperty(featureId, name);
-};
-
-ModelFeatureTable.prototype.getPropertyInherited = function (featureId, name) {
-  return this._featureTable.getProperty(featureId, name);
+  return this._propertyTable.getProperty(featureId, name);
 };
 
 ModelFeatureTable.prototype.getPropertyNames = function (results) {
-  return this._featureTable.getPropertyIds(results);
+  return this._propertyTable.getPropertyIds(results);
 };
 
 ModelFeatureTable.prototype.setProperty = function (featureId, name, value) {
-  return this._featureTable.setProperty(featureId, name, value);
+  return this._propertyTable.setProperty(featureId, name, value);
 };
 
 /**