+++ Specs/Scene/ModelExperimental/BatchTexturePipelineStageSpec.js
@@ -54,10 +54,10 @@ describe("Scene/ModelExperimental/BatchTexturePipelineStage", function () {
   it("sets up batch textures from ModelExperimental", function () {
     var renderResources = {
       shaderBuilder: new ShaderBuilder(),
-      featureTableId: "mockFeatureTable",
       model: {
-        featureTables: {
-          mockFeatureTable: {
+        featureTableId: 0,
+        featureTables: [
+          {
             featuresLength: 10,
             batchTexture: {
               batchTexture: 0,
@@ -67,44 +67,14 @@ describe("Scene/ModelExperimental/BatchTexturePipelineStage", function () {
               textureStep: 2,
             },
           },
-        },
+        ],
       },
     };
 
     BatchTexturePipelineStage.process(renderResources, {}, {});
     verifyBatchTextureShaders(renderResources.shaderBuilder);
     verifyBatchTextureUniforms(
-      renderResources.model.featureTables.mockFeatureTable,
-      renderResources.uniformMap
-    );
-  });
-
-  it("sets up batch textures from Cesium3DTileContent", function () {
-    var renderResources = {
-      shaderBuilder: new ShaderBuilder(),
-      featureTableId: "mockFeatureTable",
-      model: {
-        content: {
-          featureTables: {
-            mockFeatureTable: {
-              featuresLength: 10,
-              batchTexture: {
-                batchTexture: 0,
-                textureDimensions: {
-                  y: 2,
-                },
-                textureStep: 2,
-              },
-            },
-          },
-        },
-      },
-    };
-
-    BatchTexturePipelineStage.process(renderResources, {}, {});
-    verifyBatchTextureShaders(renderResources.shaderBuilder);
-    verifyBatchTextureUniforms(
-      renderResources.model.content.featureTables.mockFeatureTable,
+      renderResources.model.featureTables[0],
       renderResources.uniformMap
     );
   });